requirements.txt
Django>=1.8
django-debug-toolbar>=1.3.0
django-extensions>=1.5.2
ipdb>=0.8
ipython>=3.1.0
psycopg2>=2.6
Werkzeug>=0.10.4
codeforces_api==0.2a0
djangorestframework>=3.2
django-cron>=0.4.3
django-bootstrap3>=6.2.2